replace jmeter with mock in engine tests
